WiP -> Cleaned up a bunch of stuff.
